Add file extension to generated paths on POST.
This commit appends an extension to generated paths for POST requests,
reflecting the Content-Type of the request. This enables the server
to serve that file back with the correct content type.
For example, if a text/html request is posted, it will receive
the .html extension and will hence be served back as text/html.

Addresses the POST case of #275.